backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
matplotlib==3.8.2
numpy==1.26.3
stripe==7.12.0
Pillow==10.2.0
pydantic==2.5.3
python-multipart==0.0.6
